clean up and finish pr
